Extract shared email parts into partials to make it easier to reuse them later
