+++ app/views/mailer/_notification_mailer_header.html.erb
@@ -1,0 +1,30 @@
+<table width="100%" border="0" cellpadding="0" cellspacing="0" style="border-bottom: 1px solid #cccccc; margin-bottom: 30px;">
+  <tr>
+    <td style="width: 82px; height: 82px; vertical-align: top;">
+      <%= email_image_tag("logo_openproject_narrow.svg", alt: I18n.t(:'mail.logo_alt_text')) %>
+    </td>
+    <td width="100%" style="padding-top: 8px;">
+      <table>
+        <tr>
+          <td style="font-size: 24px; color: #333333; padding-bottom: 5px;">
+            <%= I18n.t(:'mail.salutation', user: @user.firstname) %>
+          </td>
+        </tr>
+        <tr>
+          <td style="font-size:14px; color: #1A67A3; font-weight: bold; padding-bottom: 10px;">
+            <%= summary %>
+          </td>
+        </tr>
+        <tr>
+          <td style="padding: 10px 0 25px 0;">
+            <a href="<%= notifications_center_url %>"
+               target="_blank"
+               style="background: #D1E5F5; padding: 8px 12px; color: #1A67A3; border: 1px solid #1A67A3; border-radius: 16px; text-decoration: none;">
+              <%= I18n.t(:'mail.notification.center') %>
+            </a>
+          </td>
+        </tr>
+      </table>
+    </td>
+  </tr>
+</table>
