+++ app/views/mailer/_notification_row.html.erb
@@ -1,0 +1,61 @@
+<a style="border: 1px solid #E0E0E0;
+  margin-bottom: 15px;
+  padding: 12px;
+  border-radius: 10px;
+  text-decoration: none;
+  display: block;"
+   href="<%= notifications_path(work_package.id) %>"
+   target="_blank">
+  <table width="100%" border="0" cellpadding="0" cellspacing="0" style="margin-bottom: 10px; font-size: 12px;">
+    <tr>
+      <td style="color: #333333;
+        background-color: #FFFFFF;
+        <%= status_colors(work_package.status) %>
+        white-space: nowrap;
+        padding: 0 5px;">
+        <%= work_package.status %>
+      </td>
+      <td style="padding-left: 5px;
+        color: #878787;">
+        #<%= work_package.id %>
+      </td>
+      <td style="white-space: nowrap;
+        padding-left: 5px;
+        color: #878787;">
+        - <%= work_package.project %>
+      </td>
+      <td style="padding-left: 5px;
+        color: #878787;
+        white-space: nowrap;">
+        <% unique_reasons = unique_reasons_of_notifications(notifications_by_work_package) %>
+        <%=  ' - ' unless unique_reasons.length === 1 && unique_reasons.first.nil? %>
+        <% unique_reasons.each_with_index do |reason, index| %>
+          <%= I18n.t(
+                :"mail.digests.work_packages.reason.#{reason || :unknown}",
+                default: '') %><%= ', ' unless unique_reasons.size-1 == index  %>
+        <% end %>
+      </td>
+      <td width="100%" style="text-align: right;">
+          <span style="background-color: #00A3FF;
+            color: white;
+            border-radius: 8px;
+            padding: 2px 7px;
+            font-size: 10px;">
+           <%= notifications_by_work_package.length %>
+          </span>
+      </td>
+    </tr>
+  </table>
+  <table width="100%" border="0" cellpadding="0" cellspacing="0" style="margin-bottom: 10px; font-size: 14px; font-weight: bold;">
+    <tr>
+      <td style="color: <%= type_color(work_package.type, '#333333') %>;">
+        <%= work_package.type.to_s.upcase %>
+      </td>
+      <td width="100%" style="padding-left: 5px; color: #333333;">
+        <%= work_package.subject %>
+      </td>
+    </tr>
+  </table>
+
+  <%= yield %>
+</a>
