+++ app/views/mailer/_notification_settings_button.html.erb
@@ -1,0 +1,5 @@
+<a href="<%= my_notifications_url %>"
+   target="_blank"
+   style="padding: 8px 12px; color: #333333; border: 1px solid #878787; border-radius: 16px; text-decoration: none; white-space: nowrap;">
+  <%= I18n.t(:'mail.notification.settings') %>
+</a>
