+++ config/locales/en.yml
@@ -1967,9 +1967,6 @@ en:
   mail:
     actions: 'Actions'
     digests:
-      center: 'Notification center'
-      salutation: 'Hey %{user}!'
-      settings: 'Change email settings'
       time_frame: "At %{time} on %{weekday}, %{date}, you have %{number_unread} unread notifications in %{number_work_packages} work packages"
       work_packages:
         comment_added: '<b>Comment</b> added'
@@ -1990,6 +1987,10 @@ en:
         updated: 'Updated'
         updated_at: '%{timestamp} ago by %{user}'
     logo_alt_text: 'OpenProject Logo'
+    notification:
+      center: 'Notification center'
+      settings: 'Change email settings'
+    salutation: 'Hey %{user}!'
 
   mail_body_account_activation_request: "A new user (%{value}) has registered. The account is pending your approval:"
   mail_body_account_information: "Your account information"
